feat: Redesign Certifications & Credentials Section
This commit redesigns the 'Certifications & Credentials' section to align with the portfolio's modern, dark-theme aesthetic, using compact glassmorphic cards and subtle animations.

Key Changes:

-   **Glassmorphism Card Design:** Replaced the previous layout with individual, compact, dark, translucent cards for each certificate, creating a more organized and visually appealing look.
-   **Updated Visual Elements:**
    -   A prominent, clean icon has been added to each card to provide a clear visual anchor.
    -   The 'View Credential' button has been restyled as a high-contrast outlined button to improve visibility and encourage interaction.
-   **Staggered Fade-Up Animation:**
    -   Integrated `framer-motion` to add a scroll-triggered animation.
    -   Each certificate card now subtly fades and slides up as it enters the viewport, with a staggered delay for a smooth, cascading effect.
-   **Code Refinements:**
    -   Created a new CSS file (`CertificatesSection.css`) to encapsulate the new styles.
    -   Refactored `CertificatesSection.jsx` to use the new card-based layout and animation components.
