+++ client/src/components/sections/CertificatesSection.jsx
@@ -1,9 +1,11 @@
 import React, { useState, useEffect } from 'react';
-import { Container, Row, Col, Button, Alert } from 'react-bootstrap';
+import { Alert, Row, Col } from 'react-bootstrap';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import { faCertificate } from '@fortawesome/free-solid-svg-icons';
 import he from 'he';
 import { getCertificates } from '../../api/apiService';
+import { motion } from 'framer-motion';
+import './CertificatesSection.css';
 
 const CertificatesSection = () => {
   const [certifications, setCertifications] = useState([]);
@@ -25,44 +27,73 @@ const CertificatesSection = () => {
     fetchCertificates();
   }, []);
 
+  const containerVariants = {
+    hidden: { opacity: 0 },
+    visible: {
+      opacity: 1,
+      transition: {
+        staggerChildren: 0.15,
+      },
+    },
+  };
+
+  const itemVariants = {
+    hidden: { y: 20, opacity: 0 },
+    visible: {
+      y: 0,
+      opacity: 1,
+      transition: {
+        duration: 0.5,
+        ease: 'easeOut',
+      },
+    },
+  };
+
   return (
-    <Container id="certificates" className="my-5 py-5">
-      <h2 className="display-5 fw-bold mb-5">Certifications & Credentials</h2>
+    <section id="certificates" className="my-5 py-5">
+      <h2 className="display-5 fw-bold mb-5 text-center">Certifications & Credentials</h2>
       {loading ? (
-        <p>Loading certifications...</p>
+        <p className="text-center">Loading certifications...</p>
       ) : error ? (
         <Alert variant="danger">{error}</Alert>
       ) : (
-        <Row>
+        <motion.div
+          as={Row}
+          xs={1}
+          md={2}
+          lg={3}
+          className="g-4"
+          variants={containerVariants}
+          initial="hidden"
+          whileInView="visible"
+          viewport={{ once: true, amount: 0.2 }}
+        >
           {certifications.map((cert) => (
-            <Col key={cert._id} md={12} className="mb-4">
-              <div className="d-flex align-items-start">
+            <motion.div as={Col} key={cert._id} variants={itemVariants}>
+              <div className="certificate-card">
                 <FontAwesomeIcon
                   icon={faCertificate}
-                  size="2x"
-                  className="text-primary me-4 mt-1"
+                  size="3x"
+                  className="certificate-icon"
                 />
-                <div>
-                  <h4 className="fw-bold">{he.decode(cert.title)}</h4>
-                  <p className="text-muted mb-2">Issued by: {he.decode(cert.issuer)}</p>
-                  {cert.url && (
-                    <Button
-                      variant="outline-primary"
-                      size="sm"
-                      href={cert.url}
-                      target="_blank"
-                      rel="noopener noreferrer"
-                    >
-                      View Credential
-                    </Button>
-                  )}
-                </div>
+                <h4 className="certificate-title">{he.decode(cert.title)}</h4>
+                <p className="certificate-issuer">Issued by: {he.decode(cert.issuer)}</p>
+                {cert.url && (
+                  <a
+                    href={cert.url}
+                    target="_blank"
+                    rel="noopener noreferrer"
+                    className="credential-button"
+                  >
+                    View Credential
+                  </a>
+                )}
               </div>
-            </Col>
+            </motion.div>
           ))}
-        </Row>
+        </motion.div>
       )}
-    </Container>
+    </section>
   );
 };
